refactor(workspaces): improve workspace management and enforce unique names
