+++ common/models/models.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Masterminds/semver"
 	"github.com/danielpaulus/go-ios/ios"
 	"github.com/danielpaulus/go-ios/ios/tunnel"
+	"github.com/google/uuid"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -149,4 +150,10 @@ type Workspace struct {
 	ID          string `json:"id" bson:"_id,omitempty"`
 	Name        string `json:"name" bson:"name"`
 	Description string `json:"description" bson:"description"`
+	IsDefault   bool   `json:"is_default" bson:"is_default"`
+}
+
+// GenerateUUID generates a new UUID for the workspace
+func (w *Workspace) GenerateUUID() {
+	w.ID = uuid.New().String()
 }
